Refactor 'note' entity into its own file
